Format string/argument fixes.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@1891 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/error.c
@@ -35,7 +35,8 @@ void print_error( const char *file, unsigned long line, const char *format, ... 
     vsnprintf( message, sizeof(message), format, ap );
     va_end (ap);
 
-    fprintf( stderr, "%s: %s\n", progname, message );
+    fprintf(stderr, "%s: %s\n", progname, message);
     if (logflag)
-	syslog( LOG_INFO, "[%u] %s:%lu:  %s", pid, file, line, message );
+	syslog(LOG_INFO, "[%lu] %s:%lu:  %s", (unsigned long)pid,
+		file, line, message );
 }
